remove warning return value

+++ base/utils.c
@@ -380,13 +380,13 @@ int my_system(char *cmd,int timeout,int *early_timeout,double *exectime,char **o
 			 */
 			(void) POPs ;
 
-			asprintf(&temp_buffer,"%s", SvPVX(ERRSV));
-
-			log_debug_info(DEBUGL_COMMANDS,0,"Embedded perl failed to compile %s, compile error %s\n",fname,temp_buffer);
-
-			logit(NSLOG_RUNTIME_WARNING,TRUE,"%s\n",temp_buffer);
-
-			my_free(temp_buffer);
+			if(asprintf(&temp_buffer,"%s", SvPVX(ERRSV))==-1)
+				logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+			else{
+				log_debug_info(DEBUGL_COMMANDS,0,"Embedded perl failed to compile %s, compile error %s\n",fname,temp_buffer);
+				logit(NSLOG_RUNTIME_WARNING,TRUE,"%s\n",temp_buffer);
+				my_free(temp_buffer);
+			}
 
 			return STATE_UNKNOWN;
 			}
@@ -403,7 +403,10 @@ int my_system(char *cmd,int timeout,int *early_timeout,double *exectime,char **o
 #endif 
 
 	/* create a pipe */
-	pipe(fd);
+	if(pipe(fd)==-1){
+		logit(NSLOG_RUNTIME_WARNING,TRUE,"Warning: pipe() in my_system() failed for command \"%s\"\n",cmd);
+		return STATE_UNKNOWN;
+		}
 
 	/* make the pipe non-blocking */
 	fcntl(fd[0],F_SETFL,O_NONBLOCK);
@@ -497,7 +500,8 @@ int my_system(char *cmd,int timeout,int *early_timeout,double *exectime,char **o
 			log_debug_info(DEBUGL_COMMANDS,0,"Embedded perl ran command %s with output %d, %s\n",fname,status,buffer);
 
 			/* write the output back to the parent process */
-			write(fd[1],buffer,strlen(buffer)+1);
+			if(write(fd[1],buffer,strlen(buffer)+1)==-1)
+				logit(NSLOG_RUNTIME_WARNING,FALSE,"Warning: Write failed. %s\n", strerror(errno));
 
 			/* close pipe for writing */
 			close(fd[1]);
@@ -507,13 +511,13 @@ int my_system(char *cmd,int timeout,int *early_timeout,double *exectime,char **o
 
 			_exit(status);
 		        }
-#endif  
+#endif
 		/******** END EMBEDDED PERL CODE EXECUTION ********/
 
 
 		/* run the command */
 		fp=(FILE *)popen(cmd,"r");
-		
+
 		/* report an error if we couldn't run the command */
 		if(fp==NULL){
 
@@ -521,7 +525,8 @@ int my_system(char *cmd,int timeout,int *early_timeout,double *exectime,char **o
 			buffer[sizeof(buffer)-1]='\x0';
 
 			/* write the error back to the parent process */
-			write(fd[1],buffer,strlen(buffer)+1);
+			if(write(fd[1],buffer,strlen(buffer)+1)==-1)
+				logit(NSLOG_RUNTIME_WARNING,FALSE,"Warning: Write failed. %s\n", strerror(errno));
 
 			result=STATE_CRITICAL;
 		        }
@@ -529,11 +534,12 @@ int my_system(char *cmd,int timeout,int *early_timeout,double *exectime,char **o
 
 			/* write all the lines of output back to the parent process */
 			while(fgets(buffer,sizeof(buffer)-1,fp))
-				write(fd[1],buffer,strlen(buffer));
+				if(write(fd[1],buffer,strlen(buffer))==-1)
+					logit(NSLOG_RUNTIME_WARNING,FALSE,"Warning: Write failed. %s\n", strerror(errno));
 
 			/* close the command and get termination status */
 			status=pclose(fp);
-			
+
 			/* report an error if we couldn't close the command */
 			if(status==-1)
 				result=STATE_CRITICAL;
@@ -778,7 +784,10 @@ int set_environment_var(char *name, char *value, int set){
 #else
 		/* needed for Solaris and systems that don't have setenv() */
 		/* this will leak memory, but in a "controlled" way, since lost memory should be freed when the child process exits */
-		asprintf(&env_string,"%s=%s",name,(value==NULL)?"":value);
+		if(asprintf(&env_string,"%s=%s",name,(value==NULL)?"":value)==-1){
+			logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+			return ERROR;
+			}
 		if(env_string)
 			putenv(env_string);
 #endif
@@ -2015,6 +2024,7 @@ int daemon_init(void){
 	char buf[256];
 	struct flock lock;
 	char *homedir=NULL;
+	int ret;
 
 #ifdef RLIMIT_CORE
 	struct rlimit limit;
@@ -2023,9 +2033,15 @@ int daemon_init(void){
 	/* change working directory. scuttle home if we're dumping core */
 	homedir=getenv("HOME");
 	if(daemon_dumps_core==TRUE && homedir!=NULL)
-		chdir(homedir);
+		ret=chdir(homedir);
 	else
-		chdir("/");
+		ret=chdir("/");
+
+	if(ret==-1){
+		logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+		cleanup();
+		exit(ERROR);
+		}
 
 	umask(S_IWGRP|S_IWOTH);
 
@@ -2102,9 +2118,12 @@ int daemon_init(void){
 
 	/* write PID to lockfile... */
 	lseek(lockfile,0,SEEK_SET);
-	ftruncate(lockfile,0);
+	if (ftruncate(lockfile,0)==-1)
+		logit(NSLOG_RUNTIME_WARNING,FALSE,"Warning: ftruncate failed %s.\n", strerror(errno));
+
 	sprintf(buf,"%d\n",(int)getpid());
-	write(lockfile,buf,strlen(buf));
+	if(write(lockfile,buf,strlen(buf))==-1)
+		logit(NSLOG_RUNTIME_WARNING,FALSE,"Warning: Write failed. %s\n", strerror(errno));
 
 	/* make sure lock file stays open while program is executing... */
 	val=fcntl(lockfile,F_GETFD,0);
@@ -2240,7 +2259,10 @@ int move_check_result_to_queue(char *checkresult_file){
 	old_umask=umask(new_umask);
 
 	/* create a safe temp file */
-	asprintf(&output_file,"%s/cXXXXXX",check_result_path);
+	if(asprintf(&output_file,"%s/cXXXXXX",check_result_path)==-1){
+		logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+		return ERROR;
+		}
 	output_file_fd=mkstemp(output_file);
 
 	/* file created okay */
@@ -2263,7 +2285,15 @@ int move_check_result_to_queue(char *checkresult_file){
 #endif
 
 		/* create an ok-to-go indicator file */
-		asprintf(&temp_buffer,"%s.ok",output_file);
+		if(asprintf(&temp_buffer,"%s.ok",output_file)==-1){
+			logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+			/* free memory */
+			my_free(output_file);
+			/* delete the original file if it couldn't be moved */
+			if(result!=0)
+			  unlink(checkresult_file);
+			return ERROR;
+		}
 		if((output_file_fd=open(temp_buffer,O_CREAT|O_WRONLY|O_TRUNC,S_IRUSR|S_IWUSR))>0)
 			close(output_file_fd);
 		my_free(temp_buffer);
@@ -2348,7 +2378,11 @@ int process_check_result_queue(char *dirname){
 			/* at this point we have a regular file... */
 
 			/* can we find the associated ok-to-go file ? */
-			asprintf(&temp_buffer,"%s.ok",file);
+			if(asprintf(&temp_buffer,"%s.ok",file)==-1){
+				logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+				closedir(dirp);
+				return ERROR;
+				}
 			result=stat(temp_buffer,&ok_stat_buf);
 			my_free(temp_buffer);
 			if(result==-1)
@@ -2554,7 +2588,10 @@ int delete_check_result_file(char *fname){
 	unlink(fname);
 
 	/* delete the ok-to-go file */
-	asprintf(&temp_buffer,"%s.ok",fname);
+	if(asprintf(&temp_buffer,"%s.ok",fname)==-1){
+		logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+		return ERROR;
+		}
 	unlink(temp_buffer);
 	my_free(temp_buffer);
 
@@ -3357,10 +3394,11 @@ int my_fcopy(char *source, char *dest){
 
 		/* open source file for reading */
 		if((source_fd=open(source,O_RDONLY,0644))>0){
-			
+
 			/* copy file contents */
 			while((bytes_read=read(source_fd,buffer,sizeof(buffer)))>0)
-				write(dest_fd,buffer,bytes_read);
+				if(write(dest_fd,buffer,bytes_read)==-1)
+					logit(NSLOG_RUNTIME_WARNING,FALSE,"Warning: Write failed. %s\n", strerror(errno));
 
 			close(source_fd);
 			close(dest_fd);
@@ -4094,7 +4132,10 @@ int submit_raw_external_command(char *cmd, time_t *ts, int *buffer_items){
 		time(&timestamp);
 
 	/* create the command string */
-	asprintf(&newcmd,"[%lu] %s",(unsigned long)timestamp,cmd);
+	if(asprintf(&newcmd,"[%lu] %s",(unsigned long)timestamp,cmd)==-1){
+		logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+		return ERROR;
+		}
 
 	/* submit the command */
 	result=submit_external_command(newcmd,buffer_items);
@@ -4484,25 +4525,48 @@ int query_update_api(void){
 	if(last_program_version==NULL || strcmp(PROGRAM_VERSION,last_program_version))
 		report_install=TRUE;
 	if(report_install==TRUE){
-		asprintf(&api_query_opts,"&firstcheck=1");
-		if(last_program_version!=NULL)
-			asprintf(&api_query_opts,"%s&last_version=%s",api_query_opts,last_program_version);
+		if(asprintf(&api_query_opts,"&firstcheck=1")==-1){
+			logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+			return ERROR;
+			}
+		if(last_program_version!=NULL && asprintf(&api_query_opts,"%s&last_version=%s",api_query_opts,last_program_version)==-1){
+			logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+			/* cleanup */
+			my_free(api_query_opts);
+			return ERROR;
+			}
 		}
 
 	/* generate the query */
-	asprintf(&api_query,"v=1&product=nagios&tinycheck=1&stableonly=1&uid=%lu",update_uid);
-	if(bare_update_check==FALSE)
-		asprintf(&api_query,"%s&version=%s%s",api_query,PROGRAM_VERSION,(api_query_opts==NULL)?"":api_query_opts);
+	if(asprintf(&api_query,"v=1&product=nagios&tinycheck=1&stableonly=1&uid=%lu",update_uid)==-1){
+			logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+			/* cleanup */
+			my_free(api_query_opts);
+			return ERROR;
+			}
+	if(bare_update_check==FALSE && asprintf(&api_query,"%s&version=%s%s",api_query,PROGRAM_VERSION,(api_query_opts==NULL)?"":api_query_opts)==-1){
+			logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+			/* cleanup */
+			my_free(api_query_opts);
+			my_free(api_query);
+			return ERROR;
+			}
 
 	/* generate the HTTP request */
-	asprintf(&buf,"POST %s HTTP/1.0\r\n",api_path);
-	asprintf(&buf,"%sUser-Agent: Nagios/%s\r\n",buf,PROGRAM_VERSION);
-	asprintf(&buf,"%sConnection: close\r\n",buf);
-	asprintf(&buf,"%sHost: %s\r\n",buf,api_server);
-	asprintf(&buf,"%sContent-Type: application/x-www-form-urlencoded\r\n",buf);
-	asprintf(&buf,"%sContent-Length: %zd\r\n",buf,strlen(api_query));
-	asprintf(&buf,"%s\r\n",buf);
-	asprintf(&buf,"%s%s\r\n",buf,api_query);
+	if(asprintf(&buf,"POST %s HTTP/1.0\r\n",api_path)==-1 ||
+	   asprintf(&buf,"%sUser-Agent: Nagios/%s\r\n",buf,PROGRAM_VERSION)==-1 ||
+	   asprintf(&buf,"%sConnection: close\r\n",buf)==-1 ||
+	   asprintf(&buf,"%sHost: %s\r\n",buf,api_server)==-1 ||
+	   asprintf(&buf,"%sContent-Type: application/x-www-form-urlencoded\r\n",buf)==-1 ||
+	   asprintf(&buf,"%sContent-Length: %zd\r\n",buf,strlen(api_query))==-1 ||
+	   asprintf(&buf,"%s\r\n",buf)==-1 ||
+	   asprintf(&buf,"%s%s\r\n",buf,api_query)==-1){
+		logit(NSLOG_RUNTIME_ERROR,FALSE,"Error: due to asprintf.\n");
+		/* cleanup */
+		my_free(api_query_opts);
+		my_free(api_query);
+		return ERROR;
+		}
 
 	/*
 	printf("SENDING...\n");
@@ -4510,7 +4574,7 @@ int query_update_api(void){
 	printf("%s",buf);
 	printf("\n");
 	*/
-	
+
 
 	result=my_tcp_connect(api_server,80,&sd,2);
 	/*printf("CONN RESULT: %d, SD: %d\n",result,sd);*/
